moved relevant components to the dashboard/sections folder
