+++ client/src/pages/dashboard/sections/ViewButtons.tsx
