+++ client/src/pages/dashboard/sections/index.ts
@@ -1,0 +1,13 @@
+import QuickStats from './QuickStats';
+import ViewButtons from './ViewButtons';
+import SearchAndFilter from './Filters';
+import ShowInfoAndExport from './ShowInfoAndExport';
+import PaginatedTable from './PaginatedTable';
+
+export {
+  QuickStats,
+  ViewButtons,
+  SearchAndFilter,
+  ShowInfoAndExport,
+  PaginatedTable,
+};
