+++ client/src/pages/dashboard/sections/paginatedTableHelpers/ComponentViewRow.tsx
