+++ client/src/pages/dashboard/sections/paginatedTableHelpers/VulnerabilityViewRow.tsx
