+++ client/src/pages/dashboard/sections/paginatedTableHelpers/index.ts
@@ -1,0 +1,11 @@
+import ComponentViewRow from './ComponentViewRow';
+import DropdownVulnerabilityViewRow from './DropdownVulnerabilityViewRow';
+import PaginationController from './PaginationController';
+import VulnerabilityViewRow from './VulnerabilityViewRow';
+
+export {
+  ComponentViewRow,
+  DropdownVulnerabilityViewRow,
+  PaginationController,
+  VulnerabilityViewRow,
+};
